requirements.txt
# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Streamlit
streamlit>=1.28.0

# HTTP requests
requests>=2.31.0

# LangChain and AI dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-google-genai>=0.0.6
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1

# Vector store
chromadb>=0.4.0

# PDF processing
pypdf>=3.17.0

# Data validation
pydantic>=2.5.0

# Environment variables
python-dotenv>=1.0.0

